Update code to reflect requested changes.
