+++ src/GBACart.cpp
@@ -772,192 +772,3 @@ void Process(GBACart::GPIO* gpio)
 
 }
 
-namespace GBACart_RumblePak
-{
-
-bool RumblePakEnabled = false;
-u16 RumbleState = 0;
-
-u16 ReadRumble(u32 addr)
-{
-    // GBATEK: For detection, AD1 seems to be pulled low when reading from it... (while) the other AD lines are open bus (containing the halfword address)...
-    // Since open-bus behavior isn't currently implemented in melonDS, we simply set all the AD lines (except for AD1) to high (at least, for now...)
-
-    return 0xFFFD;
-}
-
-void WriteRumble(u32 addr, u16 val)
-{
-    // Ported from GBE+...
-    if (((addr == 0x8000000) || (addr == 0x8001000)) && (RumbleState != val))
-    {
-        Platform::StopRumble();
-        RumbleState = val;
-        Platform::StartRumble(16);
-    }
-}
-
-}
-
-namespace GBACart_GuitarGrip
-{
-
-bool GuitarGripEnabled = false;
-u8 GuitarKeyStatus = 0x00;
-
-u8 ReadGrip8(u32 addr)
-{
-    if (addr == 0xA000000)
-    {
-        return ~GuitarKeyStatus;
-    }
-
-    return ((addr & 1) ? 0xF9 : 0xFF);
-}
-
-u16 ReadGrip16(u32 addr)
-{
-    return 0xF9FF;
-}
-
-void SetGripKey(GuitarKeys key, bool val)
-{
-    if (val == true)
-    {
-        GuitarKeyStatus |= key;
-    }
-    else
-    {
-        GuitarKeyStatus &= ~key;
-    }
-}
-
-}
-
-namespace GBACart_MemExpansionPak
-{
-
-// Code (mostly) ported from Desmume
-
-bool MemPakEnabled = false;
-
-u8 MemPakHeader[] = 
-{
-    0xFF, 0xFF, 0x96, 0x00,
-    0x00, 0x24, 0x24, 0x24,
-    0xFF, 0xFF, 0xFF, 0xFF,
-    0xFF, 0xFF, 0xFF, 0x7F
-};
-
-u8 MemPakMemory[0x800000];
-bool MemPakRAMLock = true;
-
-u8 ReadMemPak8(u32 addr)
-{
-    if ((addr >= 0x80000B0)  && (addr < 0x80000C0))
-    {
-        return *(u8*)&MemPakHeader[(addr & 0xF)];
-    }
-    else if ((addr >= 0x9000000) && (addr < 0x9800000))
-    {
-        return *(u8*)&MemPakMemory[(addr - 0x9000000)];
-    }
-		
-    return 0xFF;
-}
-
-void WriteMemPak8(u32 addr, u8 val)
-{
-    if (MemPakRAMLock == true)
-    {
-        return;
-    }
-
-    if ((addr >= 0x9000000) && (addr < 0x9800000))
-    {
-        *(u8*)&MemPakMemory[(addr - 0x9000000)] = val;
-    }
-    
-    return;
-}
-
-u16 ReadMemPak16(u32 addr)
-{
-    if ((addr >= 0x80000B0)  && (addr < 0x80000C0))
-    {
-        return *(u16*)&MemPakHeader[(addr & 0xF)];
-    }
-    else if (addr == 0x801FFFC)
-    {
-        return 0x7FFF;
-    }
-    else if (addr == 0x8240002)
-    {
-        return 0x0000;
-    }
-    else if ((addr >= 0x9000000) && (addr < 0x9800000))
-    {
-        return *(u16*)&MemPakMemory[(addr - 0x9000000)];
-    }
-
-    return 0xFFFF;
-}
-
-void WriteMemPak16(u32 addr, u16 val)
-{
-    if (addr == 0x8240000)
-    {
-        MemPakRAMLock = !(val & 0x1);
-        return;
-    }
-
-    if (MemPakRAMLock == true)
-    {
-        return;
-    }
-
-    if ((addr >= 0x9000000) && (addr < 0x9800000))
-    {
-        *(u16*)&MemPakMemory[(addr - 0x9000000)] = val;
-    }
-    
-    return;
-}
-
-u32 ReadMemPak32(u32 addr)
-{
-    if ((addr >= 0x80000B0) && (addr < 0x80000C0))
-    {
-        return *(u32*)&MemPakHeader[(addr & 0xF)];
-    }
-    else if ((addr >= 0x9000000) && (addr < 0x9800000))
-    {
-        return *(u32*)&MemPakMemory[(addr - 0x9000000)];
-    }
-    
-    return 0xFFFFFFFF;
-}
-
-void WriteMemPak32(u32 addr, u32 val)
-{
-    if (MemPakRAMLock == true)
-    {
-        return;
-    }
-
-    if ((addr >= 0x9000000) && (addr < 0x9800000))
-    {
-        *(u32*)&MemPakMemory[(addr - 0x9000000)] = val;
-    }
-    
-    return;
-}
-
-void DoSavestate(Savestate* file)
-{
-    file->Section("MemExpansionPack");
-    file->Var8((u8*)&MemPakRAMLock);
-    file->VarArray(MemPakMemory, 0x800000);
-}
-
-}
