+++ src/GBACart.h
@@ -84,56 +84,4 @@ void Process(GBACart::GPIO* gpio);
 
 }
 
-namespace GBACart_RumblePak
-{
-
-extern bool RumblePakEnabled;
-extern u16 RumbleState;
-
-u16 ReadRumble(u32 addr);
-void WriteRumble(u32 addr, u16 val);
-
-}
-
-
-enum GuitarKeys : int
-{
-    Green = 0x40,
-    Red = 0x20,
-    Yellow = 0x10,
-    Blue = 0x08,
-};
-
-namespace GBACart_GuitarGrip
-{
-
-extern bool GuitarGripEnabled;
-extern u8 GuitarKeyStatus;
-
-u8 ReadGrip8(u32 addr);
-u16 ReadGrip16(u32 addr);
-
-void SetGripKey(GuitarKeys key, bool val);
-}
-
-namespace GBACart_MemExpansionPak
-{
-
-extern bool MemPakEnabled;
-extern u8 MemPakHeader[];
-extern u8 MemPakMemory[0x800000];
-extern bool MemPakRAMLock;
-
-u8 ReadMemPak8(u32 addr);
-void WriteMemPak8(u32 addr, u8 val);
-
-u16 ReadMemPak16(u32 addr);
-void WriteMemPak16(u32 addr, u16 val);
-
-u32 ReadMemPak32(u32 addr);
-void WriteMemPak32(u32 addr, u32 val);
-
-void DoSavestate(Savestate* file);
-}
-
 #endif // GBACART_H
